--- src/contact/adapters/__tests__/contactFormAdapter.spec.ts
@@ -1,38 +1,0 @@
-import { modifyFormData } from '../contactFormAdapter';
-
-let userAgentGetter: jest.SpyInstance;
-
-const mockData = {
-  email: 'pat@pencaster.co.uk',
-  name: 'Postman Pat',
-  subject: 'General enquiry',
-  message: 'I would like this entry to be updated',
-  privacy: 'on',
-  requiredForRobots: '',
-  referrer: '/uniprotkb/P05067',
-};
-
-describe('Test contact form adapter', () => {
-  beforeEach(() => {
-    userAgentGetter = jest.spyOn(window.navigator, 'userAgent', 'get');
-  });
-
-  it('should parse the inputs and generate a generic form', () => {
-    userAgentGetter.mockReturnValue('Browser and OS info');
-    const formData1 = new FormData();
-    for (const [key, value] of Object.entries(mockData)) {
-      formData1.set(key, value);
-    }
-
-    const formData2 = modifyFormData(formData1, 'random_token');
-
-    expect(formData2.get('message')).not.toBe(mockData.message);
-    expect(formData2.get('message')).toEqual(
-      expect.stringContaining(mockData.message)
-    );
-
-    expect(formData2.has('token')).toBe(true);
-
-    expect(Object.fromEntries(formData2)).toMatchSnapshot();
-  });
-});
